+++ src/sim/dsl/Population.scala
@@ -2,13 +2,13 @@ package sim.dsl
 
 object Population {
 
-	var nbPop = 0
+	var nbPop = 0.
 
-	var students = 0
-	var middle_ages = 0
-	var teenagers = 0
+	var students = 0.
+	var middle_ages = 0.
+	var teenagers = 0.
 
-	def is_composed_by(nb: Int) {
+	def is_composed_by(nb: Double) {
 
 		nbPop += nb
